EnduroTrails.AnalizerTest/Distance/TotalDistanceAnalizerTest.cs
using System;
using System.Linq;
using EnduroTrails.Analizer.Distance;
using EnduroTrails.Analizer.Distance.Abstract;
using EnduroTrails.Analizer.Utility;
using EnduroTrails.Analizer.Utility.Abstract;
using EnduroTrails.AnalizerTest.Helper;
using EnduroTrails.FileReader;
using EnduroTrails.Model;
using Xunit;

namespace EnduroTrails.AnalizerTest.Distance
{
    public class TotalDistanceAnalizerTest
    {
        private readonly IDistanceAnalizer _totalDistanceAnalizer;
        private readonly IFileReader _fileReader;
        private const double TotalDistanceInKm = 4.400;
        private const double TotalDistanceInM = 2.2;


        public TotalDistanceAnalizerTest()
        {
            _fileReader = FileReaderContener.GetFileReader();
            IDistanceLocationsAnalizer distanceLocationsAnalizer = new DistanceLocationsAnalizer();
            _totalDistanceAnalizer = new TotalDistanceAnalizer(distanceLocationsAnalizer);
        }

        [Fact]
        public void TotalDistance()
        {            
            var locations = _fileReader.ReadWayPoints().ToArray();
            double result = _totalDistanceAnalizer.AnalizeDistance(locations);

            double resultInM = Math.Round(result * 0.8684, 4);
            double resultInKm = Math.Round(result * 1.609344, 3);
            //compare these values does not make sense !!
            //Assert.Equal(resultInKm, TotalDistanceInKm);
            //Assert.Equal(resultInM, TotalDistanceInM);
            Assert.InRange(resultInKm,4.000,TotalDistanceInKm);
            Assert.InRange(resultInM, 2.000, TotalDistanceInM);
        }
    }
}
